+++ CHANGELOG.md
@@ -3,14 +3,16 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.1/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/tanvincible/chassis/compare/v0.1.0-alpha.1...HEAD) - 18 January 2026
+## [Unreleased](https://github.com/tanvincible/chassis/compare/v0.1.0-alpha.1...HEAD) - 19 January 2026
 
 ### Merged
+- feat: fixed-width HNSW node records with O(1) [`#5`](https://github.com/tanvincible/chassis/pull/5)
 - feat: Add zero-copy vector slice access via get_vector_slice() [`#4`](https://github.com/tanvincible/chassis/pull/4)
 
 ### Added
 
-- feat: fixed-width HNSW node records with O(1) ([62babac](https://github.com/tanvincible/chassis/commit/62babacf4d2fd02fa7c6d53b971bda3e70c6563a))
+- feat: implement Graph I/O with persistent header and zero-allocation iteration ([fe46503](https://github.com/tanvincible/chassis/commit/fe46503cb6e6409c8e7a7add17150ac87ea123cc))
+- feat: add GraphHeader and graph I/O methods to HnswGraph and Storage ([fc5b465](https://github.com/tanvincible/chassis/commit/fc5b4655972b39981b696141cfce685d5cf308ee))
 
 ### Documentation